backend/requirements/dev.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.8.0
django-debug-toolbar==5.1.0
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
idna==3.10
kombu==5.5.4
oauthlib==3.2.2
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
reportlab==4.4.3
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
django-filter==25.1

